Update proj1 masahat mosalas.py

+++ proj1 masahat mosalas.py
@@ -1,7 +1,7 @@
 
 
 
-x=int(input('ghaedeh ra vard konid'))
-y=int(input('ertefaee  ra vard konid'))
+x=float(input('ghaedeh ra vard konid'))
+y=float(input('ertefaee  ra vard konid'))
 masahat=(x*y)/2
 print("masahat=",masahat)
